refactor(swagger) comments response
