+++ server/src/modules/comment/dto/index.ts
@@ -1,4 +1,5 @@
 export * from "./create.dto";
 export * from "./comment.docs";
+export * from "./reject-comment.dto";
 export * from "./confirm-comment.dto";
 export * from "./find-all-unconfirmed.dto";
